skinova_backend/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
motor==3.6.0
pymongo==4.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic[email]==2.10.3
python-dotenv==1.0.1
boto3==1.35.80
httpx==0.28.1
python-multipart==0.0.19
Pillow==11.0.0
easyocr==1.7.2
pytesseract==0.3.13
bson==0.5.10
